# swagger/output/swagger.json
{
	"basePath": "/",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"MapOfBoolean": {
			"properties": {},
			"type": "object",
			"additionalProperties": {
				"type": "boolean"
			}
		},
		"JwtToken": {
			"properties": {
				"name": {
					"type": "string",
					"description": "User's display name "
				},
				"session": {
					"type": "string"
				},
				"user": {
					"type": "string"
				},
				"scope": {
					"$ref": "#/definitions/MapOfBoolean",
					"description": "List of scope or null is all scope (*) "
				},
				"expires": {
					"type": "number",
					"format": "double",
					"description": "Valid until "
				}
			},
			"required": [
				"name",
				"session",
				"user",
				"scope",
				"expires"
			],
			"type": "object"
		},
		"RoleEntity": {
			"properties": {
				"name": {
					"type": "string"
				},
				"scope": {
					"type": "string"
				},
				"_id": {
					"type": "object"
				},
				"created": {
					"type": "number",
					"format": "double"
				},
				"updated": {
					"type": "number",
					"format": "double"
				},
				"deleted": {
					"type": "number",
					"format": "double"
				}
			},
			"required": [
				"name",
				"_id"
			],
			"type": "object"
		},
		"RoleViewWithPagination": {
			"properties": {
				"roles": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/RoleEntity"
					}
				},
				"totalItems": {
					"type": "number",
					"format": "double"
				}
			},
			"required": [
				"roles",
				"totalItems"
			],
			"type": "object"
		},
		"RoleView": {
			"properties": {
				"code": {
					"type": "string",
					"description": "role code "
				},
				"scope": {
					"type": "string",
					"description": "role scope "
				}
			},
			"required": [
				"code",
				"scope"
			],
			"type": "object"
		}
	},
	"info": {
		"title": "@gtm/service.user",
		"version": "1.0.0",
		"license": {
			"name": "UNLICENSED"
		}
	},
	"paths": {
		"/api/user/v1/system/version": {
			"post": {
				"operationId": "GetVersion",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "string"
						}
					}
				},
				"description": "Get current system version info ",
				"tags": [
					"System"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/api/user/v1/system/loggedin": {
			"get": {
				"operationId": "GetLoggedin",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"type": "boolean"
						}
					}
				},
				"description": "Check loggedin status ",
				"tags": [
					"System"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/api/user/v1/session/current": {
			"get": {
				"operationId": "GetCurrent",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/JwtToken"
						}
					}
				},
				"description": "Check current session info ",
				"tags": [
					"Session"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/api/user/v1/role": {
			"get": {
				"operationId": "GetEntities",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/RoleViewWithPagination"
						}
					}
				},
				"description": "Get Roles ",
				"tags": [
					"Role"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "query",
						"required": false,
						"type": "string"
					},
					{
						"in": "query",
						"name": "pageNumber",
						"required": false,
						"format": "double",
						"type": "number"
					},
					{
						"in": "query",
						"name": "itemCount",
						"required": false,
						"format": "double",
						"type": "number"
					}
				]
			},
			"post": {
				"operationId": "CreateEntity",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/RoleEntity"
						}
					}
				},
				"description": "Create New Role ",
				"tags": [
					"Role"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "body",
						"name": "roleView",
						"required": true,
						"schema": {
							"$ref": "#/definitions/RoleView"
						}
					}
				]
			}
		},
		"/api/user/v1/role/{id}": {
			"get": {
				"operationId": "GetEntity",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/RoleEntity"
						}
					}
				},
				"description": "Get Role by Id ",
				"tags": [
					"Role"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			},
			"put": {
				"operationId": "UpdateEntity",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/RoleEntity"
						}
					}
				},
				"description": "Update Role ",
				"tags": [
					"Role"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "roleView",
						"required": true,
						"schema": {
							"$ref": "#/definitions/RoleView"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteEntity",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Delete Role ",
				"tags": [
					"Role"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					}
				]
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {
		"jwt": {
			"type": "oauth2",
			"authorizationUrl": "/swagger/authorization/google",
			"tokenUrl": "/swagger/token/google",
			"flow": "accessCode",
			"scopes": {
				"admin": "Grants read and write access to administrative information",
				"write": "Grants write access",
				"read": "Grants read access"
			}
		}
	}
}